pointer section in C and fixes in python project

+++ Python/question7.py
@@ -1,5 +1,5 @@
 # Write a program that asks the user for a weight in kilograms and converts it to
 # pounds. There are 2.2 pounds in a kilogram.
 
-weight = int(input("Enter your weight (in kg): "))
+weight = float(input("Enter your weight (in kg): "))
 print(f"There are {weight*2.2} pounds in {weight} kilogram")
